leds: drive V2 status LEDs via native ctt-leds + gpio-led overlay
V2 status LEDs were the last in-process-hardware holdout: station-radio-interface
-> HTTP POST /led -> station-hardware-server -> onoff GPIO. Migrate them to the
same pattern as everything else — native daemon owns the hardware, Node writes a
/run/ctt contract — using the kernel gpio-led overlay.

- system: leds-overlay.sh + ctt-leds-overlay.service (board-gated, mirror
  buttons-overlay) declare the V2 LEDs as gpio-led nodes on BCM 38/39/40
  (/sys/class/leds/ctt-led-{gps,a,b}); reboot-once on change; V3 has no overlay
  (its LEDs are on the SX1509B). Added to install-systemd MUST_BE_ENABLED.
- native ctt-leds: pick a backend at startup — SX1509B (V3) or gpio-led sysfs
  (V2), idle if neither; same desired-state file and blink logic. VERSION 0.2.0
  (fleet pin stays 0.1.0 until the release is cut + validated).
- node: station-leds.js is now a single board-agnostic writer to /run/ctt/leds;
  deleted station-leds-v2.js / -v3.js / led-driver.js, the V2 GPIO driver
  hardware/led-driver/v2-driver.js, and the now-unused /led route + app.js mount.
- docs: root/native/system/hardware/station-radio-interface READMEs updated.

Cross-build clean for armhf. Hardware validation deferred to a full
V2/V3.0/V3.3 LED-mapping pass.
