+++ src/station-hardware-server/app.js
@@ -8,7 +8,6 @@ import internetRouter from './routes/internet.js'
 import peripheralRouter from './routes/peripherals.js'
 import radioRouter from './routes/radio-server.js'
 import controlRouter from './routes/control.js'
-import ledRouter from './routes/led.js'
 import programRadiosRouter from './routes/program-radios.js'
 import listDevicesRouter from './routes/list-devices.js'
 // import standByRouter from './routes/standby.js'
@@ -23,7 +22,6 @@ app.use('/modem', modemRouter)
 app.use('/sensor', sensorRouter)
 app.use('/usb', usbRouter)
 app.use('/gps', gpsRouter)
-app.use('/led', ledRouter)
 app.use('/peripherals', peripheralRouter)
 app.use('/internet', internetRouter)
 app.use('/radio', radioRouter)
