+++ system/scripts/hooks/post-merge.d/install-systemd.sh
@@ -37,6 +37,7 @@ MUST_BE_ENABLED=(
   ctt-board-detect.service         # boot-time hardware identity; writes /run/ctt/board.env
   ctt-rtc-overlay.service        # match RTC dtoverlay to detected board (early boot)
   ctt-buttons-overlay.service    # match button gpio-key overlays to board (early boot)
+  ctt-leds-overlay.service       # match LED gpio-led overlays to board (V2 GPIO LEDs; early boot)
   ctt-sensors.service            # I2C ADC + temp reader -> /run/ctt/sensors.json
   ctt-leds.service               # status LED driver (SX1509B) <- /run/ctt/leds
   ctt-lcd.service                # character LCD driver (HD44780/PCF8574) <- /run/ctt/lcd
