+++ system/systemd/ctt-leds-overlay.service
@@ -1,0 +1,23 @@
+# Ensure the LED gpio-led overlays in /boot/config.txt match the detected board
+# (V2 status LEDs are on GPIO; V3 LEDs are on the SX1509B with no overlay); fix +
+# reboot once on a mismatch. Once applied, the kernel exposes the V2 LEDs at
+# /sys/class/leds/ctt-led-* for the native ctt-leds daemon to drive.
+#
+# Ordered after ctt-board-detect (needs the board version) and before ctt-leds
+# (which actuates the resulting /sys/class/leds nodes). Reboots only on a
+# board/overlay mismatch, so on a normal boot it is a fast no-op.
+
+[Unit]
+Description=CTT LED gpio-led overlays match board
+After=ctt-board-detect.service
+Requires=ctt-board-detect.service
+Before=ctt-leds.service
+Conflicts=shutdown.target
+
+[Service]
+Type=oneshot
+RemainAfterExit=yes
+ExecStart=/lib/ctt/sensor-station-software/system/scripts/leds-overlay.sh
+
+[Install]
+WantedBy=sysinit.target
